Fix get product

+++ internal/models/products.go
@@ -8,7 +8,7 @@ import (
 
 type Product struct {
 	ID          primitive.ObjectID   `bson:"_id"`
-	Price       float32              `bson:"price"`
+	Price       float64              `bson:"price"`
 	ShopID      primitive.ObjectID   `bson:"shop_id"`
 	InventoryID primitive.ObjectID   `bson:"inventory_id"`
 	CategoryID  []primitive.ObjectID `bson:category_id`
